Remove the HTTP2 and HTTP3 implementations on browser-wasm.
This allows for HPackEncoder and QPackEncoder to be trimmed in browser wasm.

Contributes to #44534
