+++ src/libraries/System.Net.Http/src/System/Net/Http/Headers/KnownHeader.cs
@@ -2,12 +2,11 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Diagnostics;
-using System.Net.Http.HPack;
 using System.Text;
 
 namespace System.Net.Http.Headers
 {
-    internal sealed class KnownHeader
+    internal sealed partial class KnownHeader
     {
         public KnownHeader(string name, int? http2StaticTableIndex = null, int? http3StaticTableIndex = null) :
             this(name, HttpHeaderType.Custom, parser: null, knownValues: null, http2StaticTableIndex, http3StaticTableIndex)
@@ -26,13 +25,7 @@ namespace System.Net.Http.Headers
             Parser = parser;
             KnownValues = knownValues;
 
-            Http2EncodedName = http2StaticTableIndex.HasValue ?
-                HPackEncoder.EncodeLiteralHeaderFieldWithoutIndexingToAllocatedArray(http2StaticTableIndex.GetValueOrDefault()) :
-                HPackEncoder.EncodeLiteralHeaderFieldWithoutIndexingNewNameToAllocatedArray(name);
-
-            Http3EncodedName = http3StaticTableIndex.HasValue ?
-                QPack.QPackEncoder.EncodeLiteralHeaderFieldWithStaticNameReferenceToArray(http3StaticTableIndex.GetValueOrDefault()) :
-                QPack.QPackEncoder.EncodeLiteralHeaderFieldWithoutNameReferenceToArray(name);
+            Initialize(http2StaticTableIndex, http3StaticTableIndex);
 
             var asciiBytesWithColonSpace = new byte[name.Length + 2]; // + 2 for ':' and ' '
             int asciiBytes = Encoding.ASCII.GetBytes(name, asciiBytesWithColonSpace);
@@ -41,6 +34,8 @@ namespace System.Net.Http.Headers
             asciiBytesWithColonSpace[asciiBytesWithColonSpace.Length - 1] = (byte)' ';
             AsciiBytesWithColonSpace = asciiBytesWithColonSpace;
         }
+
+        partial void Initialize(int? http2StaticTableIndex, int? http3StaticTableIndex);
 
         public string Name { get; }
         public HttpHeaderParser? Parser { get; }
@@ -52,7 +47,5 @@ namespace System.Net.Http.Headers
         public string[]? KnownValues { get; }
         public byte[] AsciiBytesWithColonSpace { get; }
         public HeaderDescriptor Descriptor => new HeaderDescriptor(this);
-        public byte[] Http2EncodedName { get; }
-        public byte[] Http3EncodedName { get; }
     }
 }
