+++ src/libraries/System.Net.Http/src/System/Net/Http/HttpMethod.cs
@@ -1,20 +1,17 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Collections.Generic;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Net.Http.QPack;
-using System.Threading;
 
 namespace System.Net.Http
 {
-    public class HttpMethod : IEquatable<HttpMethod>
+    public partial class HttpMethod : IEquatable<HttpMethod>
     {
         private readonly string _method;
         private readonly int? _http3Index;
 
-        private byte[]? _http3EncodedBytes;
         private int _hashcode;
 
         private static readonly HttpMethod s_getMethod = new HttpMethod("GET", http3StaticTableIndex: H3StaticTable.MethodGet);
@@ -78,20 +75,6 @@ namespace System.Net.Http
         public string Method
         {
             get { return _method; }
-        }
-
-        internal byte[] Http3EncodedBytes
-        {
-            get {
-                byte[]? http3EncodedBytes = Volatile.Read(ref _http3EncodedBytes);
-                if (http3EncodedBytes is null) {
-                    Volatile.Write (ref _http3EncodedBytes, http3EncodedBytes = _http3Index is int index && index >= 0 ?
-                        QPackEncoder.EncodeStaticIndexedHeaderFieldToArray(index) :
-                        QPackEncoder.EncodeLiteralHeaderFieldWithStaticNameReferenceToArray(H3StaticTable.MethodGet, _method));
-                }
-
-                return http3EncodedBytes;
-            }
         }
 
         public HttpMethod(string method)
